requirements.txt
numpy==1.26.0
flax==0.8.2
optax==0.2.1
einops
pyyaml